requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
beanie==1.24.0
motor==3.3.2
pymongo==4.6.1

# Data Validation
pydantic==2.5.3
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
bcrypt==4.1.2

# Environment & Configuration
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0

# Caching (Optional)
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Production Server
gunicorn==21.2.0
